Refactored CommonReposiotry to take clazz in constructor.
