+++ src/main/java/com/danigu/blog/ServerApplication.java
@@ -1,7 +1,6 @@
 package com.danigu.blog;
 
 import com.danigu.blog.dic.StationLocator;
-import com.danigu.blog.servlet.BlogServlet;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.servlet.ServletContextHandler;
 import org.eclipse.jetty.servlet.ServletHolder;
