+++ src/main/java/com/danigu/blog/comment/service/CommentTransformer.java
@@ -3,15 +3,15 @@ package com.danigu.blog.comment.service;
 import com.danigu.blog.comment.Comment;
 import com.danigu.blog.comment.CommentEntity;
 import com.danigu.blog.common.service.Transformer;
-import com.danigu.blog.post.service.PostEntityDTOTransformer;
+import com.danigu.blog.post.service.PostTransformer;
 import lombok.AllArgsConstructor;
 
 /**
  * @author dani
  */
 @AllArgsConstructor
-public class CommentEntityDTOTransformer implements Transformer<CommentEntity, Comment> {
-    private final PostEntityDTOTransformer postTransformer;
+public class CommentTransformer implements Transformer<CommentEntity, Comment> {
+    private final PostTransformer postTransformer;
 
     public Comment toEntity(CommentEntity entity) {
         return new CommentDTO(entity.getId(), postTransformer.toEntity(entity.getPost()), entity.getContent());
