+++ src/main/java/com/danigu/blog/post/service/PostTransformer.java
@@ -8,7 +8,7 @@ import com.danigu.blog.post.PostEntity;
  * Converts between the implementation of PostDTO in the service and the DTO of persistence.
  * @author dani
  */
-public class PostEntityDTOTransformer implements Transformer<PostEntity, Post> {
+public class PostTransformer implements Transformer<PostEntity, Post> {
     public Post toEntity(PostEntity entity) {
         return new PostDTO(entity.getId(), entity.getName(), entity.getContent());
     }
